Move strip_dxil_validator_path to getLocalConfig to cover FileCheck test too.
